requirements.txt
confuse==1.4.0
et-xmlfile==1.0.1
joblib==1.0.1
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
